Restposten vom IOWarrior entfernt
daySettingArray mit [8][16]
neu:
- (uint8_t)freeDaySettingline; // freie Zeile in daySettingArray suchen

- (NSArray*)daySettingDataVon:(uint8_t)raum vonObjekt:(uint8_t)objekt anWochentag:(uint8_t)wochentag; // data von bestimmtem Objekt abfragen sofern vorhanden

neu:  [tempWochenplanDic setObject:[NSNumber numberWithInt:k] forKey:@"wt"]; // als zahl eingefügt

WriteStandardAktion:
AVRClient l 970ff :

Sicherung der neuen Daten in daySettingArray

+++ WebInterface_210_0/Source/rMehrkanalDiagramm.m
@@ -247,7 +247,7 @@ StartwertX=x;
 }
 
 
-- (void)setOffsetY:(float)y
+- (void)setOffsetY:(int)y
 {
